# app/Modules/Platform/Resources/PlatformTenantResource.php
<?php

namespace App\Modules\Platform\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

class PlatformTenantResource extends JsonResource
{
    public function toArray(Request $request): array
    {
        return [
            'public_id' => $this->public_id,
            'name_ar' => $this->name_ar,
            'name_en' => $this->name_en ?? $this->name_ar,
            'slug' => $this->slug,
            'domain' => $this->domain,
            'database_name' => $this->database_name,
            'logo_path' => $this->logo_path,
            'default_language' => $this->default_language,
            'timezone' => $this->timezone,
            'is_active' => $this->is_active,
            'settings' => $this->settings,
            'created_at' => $this->created_at?->toIso8601String(),
            'updated_at' => $this->updated_at?->toIso8601String(),
        ];
    }
}