# openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://api.momentum.test/v1",
            "description": "Local"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/organization/departments/{department}/reactivate": {
            "post": {
                "operationId": "department.reactivate",
                "tags": [
                    "Department"
                ],
                "parameters": [
                    {
                        "name": "department",
                        "in": "path",
                        "required": true,
                        "description": "The department public id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Tenant",
                        "in": "header",
                        "required": true,
                        "description": "Tenant slug or public ID for multi-tenant resolution.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`DepartmentResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepartmentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/platform/admins/{admin}/reactivate": {
            "post": {
                "operationId": "platformAdmin.reactivate",
                "tags": [
                    "PlatformAdmin"
                ],
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "required": true,
                        "description": "The admin public id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Tenant",
                        "in": "header",
                        "required": true,
                        "description": "Tenant slug or public ID for multi-tenant resolution.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PlatformAdminResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlatformAdminResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/platform/tenants/{tenant}/suspend": {
            "post": {
                "operationId": "platformTenant.suspend",
                "tags": [
                    "PlatformTenant"
                ],
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "description": "The tenant public id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Tenant",
                        "in": "header",
                        "required": true,
                        "description": "Tenant slug or public ID for multi-tenant resolution.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PlatformTenantResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlatformTenantResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/platform/tenants/{tenant}/reactivate": {
            "post": {
                "operationId": "platformTenant.reactivate",
                "tags": [
                    "PlatformTenant"
                ],
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "description": "The tenant public id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Tenant",
                        "in": "header",
                        "required": true,
                        "description": "Tenant slug or public ID for multi-tenant resolution.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PlatformTenantResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlatformTenantResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/platform/tenants/{tenant}/run-migrations": {
            "post": {
                "operationId": "platformTenant.runMigrations",
                "tags": [
                    "PlatformTenant"
                ],
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "description": "The tenant public id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Tenant",
                        "in": "header",
                        "required": true,
                        "description": "Tenant slug or public ID for multi-tenant resolution.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Migrations dispatched. Check audit events for status."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/organization/positions/{position}/reactivate": {
            "post": {
                "operationId": "position.reactivate",
                "tags": [
                    "Position"
                ],
                "parameters": [
                    {
                        "name": "position",
                        "in": "path",
                        "required": true,
                        "description": "The position public id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Tenant",
                        "in": "header",
                        "required": true,
                        "description": "Tenant slug or public ID for multi-tenant resolution.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PositionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PositionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/iam/users/{user}/reactivate": {
            "post": {
                "operationId": "user.reactivate",
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "description": "The user public id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Tenant",
                        "in": "header",
                        "required": true,
                        "description": "Tenant slug or public ID for multi-tenant resolution.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`UserResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "DepartmentResource": {
                "type": "object",
                "properties": {
                    "public_id": {
                        "type": "string"
                    },
                    "parent_department_id": {
                        "type": "string"
                    },
                    "name_ar": {
                        "type": "string"
                    },
                    "name_en": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "string"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DepartmentResource"
                        }
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "public_id",
                    "parent_department_id",
                    "name_ar",
                    "name_en",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "DepartmentResource"
            },
...
            "PlatformAdminResource": {
                "type": "object",
                "properties": {
                    "public_id": {
                        "type": "string"
                    },
                    "name_ar": {
                        "type": "string"
                    },
                    "name_en": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "account_type": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "public_id",
                    "name_ar",
                    "name_en",
                    "email",
                    "account_type",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "PlatformAdminResource"
            },
...
            "PlatformTenantResource": {
                "type": "object",
                "properties": {
                    "public_id": {
                        "type": "string"
                    },
                    "name_ar": {
                        "type": "string"
                    },
                    "name_en": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "domain": {
                        "type": "string"
                    },
                    "database_name": {
                        "type": "string"
                    },
                    "logo_path": {
                        "type": "string"
                    },
                    "default_language": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "string"
                    },
                    "settings": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "public_id",
                    "name_ar",
                    "name_en",
                    "slug",
                    "domain",
                    "database_name",
                    "logo_path",
                    "default_language",
                    "timezone",
                    "is_active",
                    "settings",
                    "created_at",
                    "updated_at"
                ],
                "title": "PlatformTenantResource"
            },
...
            "PositionResource": {
                "type": "object",
                "properties": {
                    "public_id": {
                        "type": "string"
                    },
                    "department": {
                        "type": "object",
                        "properties": {
                            "public_id": {
                                "type": "string"
                            },
                            "name_ar": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "public_id",
                            "name_ar"
                        ]
                    },
                    "title_ar": {
                        "type": "string"
                    },
                    "title_en": {
                        "type": "string"
                    },
                    "reports_to_position_id": {
                        "type": "string"
                    },
                    "authority_grade": {
                        "type": "object",
                        "properties": {
                            "public_id": {
                                "type": "string"
                            },
                            "rank": {
                                "type": "string"
                            },
                            "name_ar": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "public_id",
                            "rank",
                            "name_ar"
                        ]
                    },
                    "is_department_head": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "string"
                    },
                    "current_occupant": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "properties": {
                            "public_id": {
                                "type": "string"
                            },
                            "name_ar": {
                                "type": "string"
                            },
                            "name_en": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "public_id",
                            "name_ar",
                            "name_en"
                        ]
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "public_id",
                    "department",
                    "title_ar",
                    "title_en",
                    "reports_to_position_id",
                    "authority_grade",
                    "is_department_head",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "PositionResource"
            },
            "PreferredLanguage": {
                "type": "integer",
                "enum": [
                    1,
                    2
                ],
                "title": "PreferredLanguage"
            },
...
            "UserResource": {
                "type": "object",
                "properties": {
                    "public_id": {
                        "type": "string"
                    },
                    "name_ar": {
                        "type": "string"
                    },
                    "name_en": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "employee_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "account_type": {
                        "type": "integer"
                    },
                    "preferred_language": {
                        "$ref": "#/components/schemas/PreferredLanguage"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "is_out_of_office": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "public_id",
                    "name_ar",
                    "name_en",
                    "email",
                    "mobile",
                    "employee_id",
                    "account_type",
                    "preferred_language",
                    "is_active",
                    "is_out_of_office",
                    "created_at",
                    "updated_at"
                ],
                "title": "UserResource"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}